requirements/local.txt
# Inherits from tests.txt + development utilities

-r tests.txt

django-debug-toolbar==3.2.1
ipython==7.25.0